requirements.txt
fastapi
uvicorn
groq
python-dotenv
sentence-transformers
faiss-cpu
numpy<2.0
pandas>=1.3.0
numexpr>=2.8.4
bottleneck>=1.3.6